+++ src/App.js
@@ -1,6 +1,7 @@
 import React from "react";
 import "./App.css";
-import ModalLogout from "./components/modalLogout/modalLogout"
+import ModalLogout from "./components/modalLogout/modalLogout.jsx"
+
 // console.log(moment().format());
 
 function App() {
