+++ src/components/modalLogout/modalLogout.module.css
@@ -1,10 +1,34 @@
+.centerCase {
+    display: flex;
+    justify-content: center;
+    margin-top: 50px;
+}
+
 .moduleCase {
     width: 320px;
+    position: relative;
     display: flex;
     flex-direction: column;
     text-align: center;
     align-items: center;
+    border: solid 2px #999daa;
+    justify-content: space-around;
+}
 
+.btnClose {
+    width: 24px;
+    position: absolute;
+    right: 15px;
+    top: 10px;
+}
+
+.text {
+    padding-top: 39px;
+    margin-bottom: 48px;
+    color: #212121;
+    font-size: 22px;
+    font-weight: 600;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
 }
 
 .line {
@@ -13,17 +37,76 @@
     color: #284060;
     height: 2px;
     width: 256px;
+    margin-bottom: 24px;
 }
 
 .btnCase {
     display: flex;
     flex-direction: column;
+    margin-bottom: 24px;
 }
 
 .btnNo {
-
+    outline: none;
+    padding-top: 14px;
+    padding-bottom: 14px;
+    padding-left: 28px;
+    padding-right: 28px;
+    background-color: #fc842c;
+    border-radius: 50px;
+    color: white;
+    text-shadow:  0 0 01em;
+    box-shadow: 0 0 10px orange;
 }
 
 .btnYes {
+    outline: none;
+    padding-top: 14px;
+    padding-bottom: 14px;
+    padding-left: 28px;
+    padding-right: 28px;
+    margin-bottom: 24px;
+    background-color: #fc842c;
+    border-radius: 50px;
+    color: white;
+    text-shadow:  0 0 01em;
+    box-shadow: 0 0 10px orange;
+}
 
+@media(min-width:768px){
+    .moduleCase {
+        width: 500px;
+    }
+
+    .btnClose {
+
+    }
+
+    .text {
+        margin-bottom: 24px;
+    }
+
+    .line {
+        width: 356px;
+    }
+
+    .btnCase {
+        display: flex;
+        flex-direction: row;
+        
+    }
+
+    .btnYes {
+        width: 180px;
+        margin-right: 5px;
+        margin-bottom: 0;
+    }
+
+    .btnNo {
+        margin-left: 5px;
+    }
+}
+
+@media(min-width:1024px){
+    
 }
